Naively places armies

+++ src/model/city.rs
@@ -14,7 +14,7 @@ pub struct City {
     pub location: Location,
     pub name: String,
     pub size: u8,
-    pub armies: u8,
+    pub armies: u32,
     pub owner: Option<Entity>,
     pub node: NodeIndex,
 }
